9.google-agent-sdk-gcp-assistant/requirements.txt
google-adk>=0.1.0
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
google-auth
requests
